Block saving

+++ src/services/lessonService.ts
@@ -5,8 +5,21 @@ import type { Lesson, LessonBlock } from "./types";
 // returns the lessons if successful
 // throws an error otherwise
 const getLessons = async (moduleId: number) => {
+    const token: string | null = localStorage.getItem("token");
+    if (token === null) {
+        throw new Error(`an error occurred`);
+    }
+
     const url = new URL(`${apiBaseUrl}/v1/module/${moduleId}/lessons`);
-    const response = await fetch(url, { method: "GET" });
+    const response = await fetch(
+        url, 
+        { 
+            method: "GET",
+            headers: {
+                "Authorization": `Bearer ${token}`
+            }
+        }
+    );
     if (!response.ok) {
         throw new Error(`an error occurred`);
     }
@@ -23,12 +36,18 @@ const addLesson = async (
     moduleId: number,
     title: string,
 ) => {
-    const url = `${apiBaseUrl}/v1/module/lesson`;
+    const token: string | null = localStorage.getItem("token");
+    if (token === null) {
+        throw new Error(`an error occurred`);
+    }
+
+    const url = `${apiBaseUrl}/v1/module/lesson/`;
 
     const response = await fetch(url, {
         method: "POST",
         headers: {
             "Content-Type": "application/x-www-form-urlencoded",
+            "Authorization": `Bearer ${token}`
         },
         body: new URLSearchParams({
             lesson_id: lessonId.toString(),
@@ -47,15 +66,23 @@ const addLesson = async (
 // passes the lesson_id as formdata
 // returns nothing if successful
 // throws an error containing the relevant message otherwise
-const deleteLesson = async (lessonId: number) => {
+const deleteLesson = async (moduleId: number, lessonId: number) => {
+    const token: string | null = localStorage.getItem("token");
+    if (token === null) {
+        throw new Error(`an error occurred`);
+    }
     const url = `${apiBaseUrl}/v1/module/lesson/`;
 
     const formData = new FormData();
+    formData.append("module_id", moduleId.toString());
     formData.append('lesson_id', lessonId.toString());
 
     const response = await fetch(url, {
         method: "DELETE",
-        body: formData,
+        headers: {
+            "Authorization": `Bearer ${token}`
+        },
+        body: formData
     });
     if (!response.ok) {
         const data: { message: string } = await response.json();
@@ -66,16 +93,27 @@ const deleteLesson = async (lessonId: number) => {
 // makes a GET request to the endpoint at /v1/module/lesson/{lessonId}/block
 // returns the lesson blocks if successful
 // throws an error containing the relevant message otherwise
-const getLessonBlocks = async (lessonId: number) => {
-    const url = `${apiBaseUrl}/v1/module/lesson/${lessonId}/block`;
+const getLessonBlocks = async (moduleId: number, lessonId: number) => {
+    const token: string | null = localStorage.getItem("token");
+    if (token === null) {
+        throw new Error(`an error occurred`);
+    }
+    console.log(moduleId, lessonId)
+
+    const url = `${apiBaseUrl}/v1/module/lesson/${lessonId}/block?module_id=${moduleId}`;
 
     const response = await fetch(url, {
         method: "GET",
+        headers: {
+            "Authorization": `Bearer ${token}`
+        }
     });
+
     if (!response.ok) {
         const data: { message: string } = await response.json();
         throw new Error(data.message);
     }
+
     const data: { blocks: LessonBlock[] } = await response.json();
     return data;
 };
@@ -84,15 +122,21 @@ const getLessonBlocks = async (lessonId: number) => {
 // passes the block_id, block_type, order, and data as formdata
 // returns nothing if successful
 // throws an error containing the relevant message otherwise
-const addLessonBlock = async (lessonId: number, block: LessonBlock) => {
+const addLessonBlock = async (moduleId: number, lessonId: number, block: LessonBlock) => {
+    const token: string | null = localStorage.getItem("token");
+    if (token === null) {
+        throw new Error(`an error occurred`);
+    }
     const url = `${apiBaseUrl}/v1/module/lesson/${lessonId}/block`;
 
     const response = await fetch(url, {
         method: "POST",
         headers: {
-            "Content-Type": "application/x-www-form-urlencoded"
+            "Content-Type": "application/x-www-form-urlencoded",
+            "Authorization": `Bearer ${token}`
         },
         body: new URLSearchParams({
+            module_id: moduleId.toString(),
             lesson_id: lessonId.toString(),
             block_id: block.block_id.toString(),
             block_type: block.block_type.toString(),
@@ -110,14 +154,22 @@ const addLessonBlock = async (lessonId: number, block: LessonBlock) => {
 // makes a DELETE request to the endpoint at /v1/module/lesson/{lessonId}/block
 // returns nothing if successful
 // throws an error containing the relevant message otherwise
-const deleteLessonBlock = async (lessonId: number, blockId: number) => {
+const deleteLessonBlock = async (moduleId: number, lessonId: number, blockId: number) => {
+    const token: string | null = localStorage.getItem("token");
+    if (token === null) {
+        throw new Error(`an error occurred`);
+    }
     const url = `${apiBaseUrl}/v1/module/lesson/${lessonId}/block`;
 
     const formData = new FormData();
     formData.append('block_id', blockId.toString());
+    formData.append("module_id", moduleId.toString());
 
     const response = await fetch(url, {
         method: "DELETE",
+        headers: {
+            "Authorization": `Bearer ${token}`
+        },
         body: formData,
     });
     if (!response.ok) {
